+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 `Constraint.NullableDateTime.Predicate` and `Constraint.NullableDateTime.UpTo` for filtering nullable `DateTime`
 relation keys.
 
+### Fixed
+
+Crash when using OnDiskMemoryMappedFileCollection with BTreeKeyValueDB.
+
 ## 33.6.9
 
 ### Backported